댓글 관리: 댓글 추가시 로그인 권한 기능 구현

+++ articles/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from .models import Article, Comment
 from .forms import NewPostForm, NewCommentForm
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 
 # Create your views here.
 def index(request):
@@ -38,20 +40,27 @@ def create(request):
     }
     return render(request, 'articles/create.html', context)
 
+@login_required
 def comment_create(request, article_pk):
     article = Article.objects.get(pk=article_pk)
     comment_form = NewCommentForm(request.POST)
-    print(dir(request))
     if comment_form.is_valid():
         # 이 시점에는 모델폼의 인스턴스, 객체를 반환
         comment = comment_form.save(commit=False)
         # 이 시점에서는 모델의 인스턴스를 반환
         comment.article_id = article.pk
         comment.save()
-        print(comment)
+    # 요청이 GET으로 들어오든 POST로 들어오든 게시물 상세페이지로 리디렉트
+    # GET으로 들어온다면 46번째 코드까지는 실행(유효성 검사 NOT NULL 탈락)
     return redirect('articles:detail', article.pk)
 
+@login_required
 def comment_delete(request, article_pk, comment_pk):
-    comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+    # 요청한 유저와 게시글 작성자의 pk가 일치해야 한다.
+    article = Article.objects.get(pk=article_pk)
+    if request.user == article.user:
+        comment = Comment.objects.get(pk=comment_pk)
+        comment.delete()
+        return redirect('articles:detail', article_pk)
+    else:
+        return HttpResponseForbidden()
